Revert "Fixes #1280 - react-clay-charts"
