--- packages/clay-charts-react/demo/index.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<body>
-	<div id="root"></div>
-</body>
-
-</html>
