--- packages/clay-charts-react/package.json
@@ -1,48 +1,0 @@
-{
-	"name": "clay-charts-react",
-	"version": "0.0.1",
-	"description": "Implementation of clay-charts in React.js",
-	"main": "lib/index.js",
-	"jsnext:main": "src/index.js",
-	"scripts": {
-		"compile": "./node_modules/.bin/babel -d lib/ src/ -s --ignore src/__tests__",
-		"prepublish": "npm run compile && npm run copyAssets",
-		"start": "webpack-dev-server --mode development",
-		"copyAssets": "ncp ./node_modules/clay-charts-shared/lib/css lib/css"
-	},
-	"author": "Bryce Osterhaus",
-	"license": "BSD",
-	"dependencies": {
-		"clay-charts-shared": "^0.0.1",
-		"react-billboardjs": "^1.4.3"
-	},
-	"peerDependencies": {
-		"react": "^16.5.2",
-		"react-dom": "^16.6.0"
-	},
-	"devDependencies": {
-		"@babel/cli": "^7.1.2",
-		"@babel/core": "^7.1.2",
-		"@babel/polyfill": "^7.0.0",
-		"@babel/preset-env": "^7.1.0",
-		"@babel/preset-react": "^7.0.0",
-		"babel-loader": "^8.0.4",
-		"css-loader": "^1.0.0",
-		"eslint-plugin-react": "^7.11.1",
-		"html-webpack-plugin": "^3.2.0",
-		"ncp": "^2.0.0",
-		"react": "^16.6.0",
-		"react-billboardjs": "^1.4.3",
-		"react-dom": "^16.6.0",
-		"style-loader": "^0.23.1",
-		"webpack": "^4.22.0",
-		"webpack-cli": "^3.1.2",
-		"webpack-dev-server": "^3.1.10"
-	},
-	"babel": {
-		"presets": [
-			"@babel/preset-env",
-			"@babel/preset-react"
-		]
-	}
-}
