--- packages/clay-charts-react/webpack.config.js
@@ -1,20 +1,0 @@
-const path = require('path');
-const HWP = require('html-webpack-plugin');
-
-module.exports = {
-	entry: path.join(__dirname, './demo/App.js'),
-	module: {
-		rules: [
-			{
-				test: /\.js$/,
-				exclude: /node_modules/,
-				loader: 'babel-loader',
-			},
-			{
-				test: /\.css$/,
-				use: ['style-loader', 'css-loader'],
-			},
-		],
-	},
-	plugins: [new HWP({template: path.join(__dirname, './demo/index.html')})],
-};
