--- packages/clay-charts-shared/README.md
@@ -1,20 +1,0 @@
-# clay-charts-shared
-
-Shared config between [clay-charts-react](../clay-charts-react) and [clay-charts](../clay-charts)
-
-## Setup
-
-1. Install NodeJS >= v0.12.0 and NPM >= v3.0.0, if you don't have it yet. You
-   can find it [here](https://nodejs.org).
-
-2. Install local dependencies:
-
-```
-npm install
-```
-
-3. Build the code:
-
-```
-npm run compile
-```
