--- packages/clay-charts-shared/package.json
@@ -1,26 +1,0 @@
-{
-	"name": "clay-charts-shared",
-	"version": "0.0.1",
-	"description": "",
-	"main": "lib/all.js",
-	"jsnext:main": "src/all.js",
-	"scripts": {
-		"build": "npm run sass && npm run compile && npm run copyAssets",
-		"compile": "./node_modules/.bin/babel -d lib/ src/ -s",
-		"prepublish": "npm run build",
-		"sass": "node-sass src/scss/main.scss -o lib/css --source-map true",
-		"copyAssets": "ncp src/svg lib/svg"
-	},
-	"author": "Bryce Osterhaus",
-	"license": "BSD",
-	"dependencies": {
-		"d3": "^5.7.0",
-		"metal": "^2.16.6"
-	},
-	"devDependencies": {
-		"@babel/cli": "^7.1.2",
-		"@babel/core": "^7.1.2",
-		"ncp": "^2.0.0",
-		"node-sass": "^4.9.4"
-	}
-}
