--- packages/clay-charts-shared/src/all.js
@@ -1,5 +1,0 @@
-import * as config from './config';
-import * as data from './data';
-import * as geomap from './geomap';
-
-export {config, data, geomap};
