+++ packages/clay-charts/demos/colors.html
@@ -1,13 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Demo: Clay Charts</title>
 
 	<link rel="stylesheet" href="../../../node_modules/clay-css/lib/css/atlas.css">
-	<link rel="stylesheet" href="../build/main.css">
+	<link rel="stylesheet" href="../build/clay-charts.css">
 
 	<script src="../build/globals/clay-charts.js"></script>
 
@@ -54,7 +53,6 @@
 
 
 </head>
-
 <body role="main">
 	<h1>Clay Charts - Color Palette</h1>
 
@@ -117,7 +115,7 @@
 		];
 
 		var typeSelect = document.querySelector('.chart-selector');
-		typeSelect.addEventListener('change', function (event) {
+		typeSelect.addEventListener('change', function(event) {
 			chart.type = typeSelect.value;
 		});
 
@@ -127,5 +125,4 @@
 		}, '#chart');
 	</script>
 </body>
-
 </html>
