+++ packages/clay-charts/package.json
@@ -1,63 +1,68 @@
 {
-	"name": "clay-charts",
-	"version": "2.3.4",
-	"description": "Metal.js wrapper for D3 and billboard.js",
-	"license": "BSD",
-	"repository": "https://github.com/liferay/clay/tree/master/packages/clay-charts",
-	"engines": {
-		"node": ">=0.12.0",
-		"npm": ">=3.0.0"
-	},
-	"main": "lib/charts.js",
-	"jsnext:main": "src/charts.js",
-	"files": [
-		"lib",
-		"src",
-		"test"
-	],
-	"scripts": {
-		"build": "npm run soy && webpack && ncp ./node_modules/clay-charts-shared/lib/css build",
-		"compile": "babel -d lib/ src/ -s --ignore src/__tests__",
-		"prepublish": "npm run soy && npm run compile && npm run copyAssets",
-		"link": "npm link",
-		"soy": "metalsoy --soyDeps node_modules/metal*/src/**/*.soy node_modules/clay*/src/**/*.soy",
-		"copyAssets": "ncp src/svg lib/svg && ncp ./node_modules/clay-charts-shared/lib/css lib/css"
-	},
-	"keywords": [
-		"metal"
-	],
-	"babel": {
-		"presets": [
-			"metal-jsx",
-			"es2015"
-		]
-	},
-	"dependencies": {
-		"billboard.js": "1.5.1",
-		"clay-charts-shared": "^0.0.1",
-		"d3": "5.0.0",
-		"metal": "^2.16.2",
-		"metal-component": "^2.13.2",
-		"metal-jsx": "^2.13.2",
-		"metal-soy": "^2.16.0",
-		"metal-state": "^2.16.0"
-	},
-	"devDependencies": {
-		"babel-cli": "^6.24.1",
-		"babel-core": "^6.25.0",
-		"babel-loader": "^7.0.0",
-		"babel-plugin-transform-node-env-inline": "^0.1.1",
-		"babel-preset-env": "^1.6.0",
-		"babel-preset-metal": "^4.0.0",
-		"babel-preset-metal-jsx": "^0.1.1",
-		"browserslist-config-clay": "^2.3.4",
-		"clay-css": "^2.3.4",
-		"metal-tools-soy": "^6.0.0",
-		"ncp": "^2.0.0",
-		"webpack": "^3.0.0",
-		"webpack-config-clay": "^2.3.4"
-	},
-	"browserslist": [
-		"extends browserslist-config-clay"
-	]
+  "name": "clay-charts",
+  "version": "2.3.4",
+  "description": "Metal.js wrapper for D3 and billboard.js",
+  "license": "BSD",
+  "repository": "https://github.com/liferay/clay/tree/master/packages/clay-charts",
+  "engines": {
+    "node": ">=0.12.0",
+    "npm": ">=3.0.0"
+  },
+  "main": "lib/charts.js",
+  "jsnext:main": "src/charts.js",
+  "files": [
+    "lib",
+    "src",
+    "test"
+  ],
+  "scripts": {
+    "build": "npm run soy && webpack",
+    "compile": "babel -d lib/ src/ -s --ignore src/__tests__",
+    "prepublish": "npm run soy && npm run sass && npm run compile && npm run copyAssets",
+    "link": "npm link",
+    "sass": "node-sass src/scss/main.scss -o lib/css --source-map true",
+    "soy": "metalsoy --soyDeps node_modules/metal*/src/**/*.soy node_modules/clay*/src/**/*.soy",
+    "copyAssets": "ncp src/svg lib/svg"
+  },
+  "keywords": [
+    "metal"
+  ],
+  "babel": {
+    "presets": [
+      "metal-jsx",
+      "es2015"
+    ]
+  },
+  "dependencies": {
+    "billboard.js": "1.5.1",
+    "d3": "5.0.0",
+    "metal": "^2.16.2",
+    "metal-component": "^2.13.2",
+    "metal-jsx": "^2.13.2",
+    "metal-soy": "^2.16.0",
+    "metal-state": "^2.16.0"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.0.0",
+    "babel-plugin-transform-node-env-inline": "^0.1.1",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-metal": "^4.0.0",
+    "babel-preset-metal-jsx": "^0.1.1",
+    "browserslist-config-clay": "^2.3.4",
+    "clay-css": "^2.3.4",
+    "css-loader": "^0.28.4",
+    "extract-text-webpack-plugin": "^3.0.1",
+    "metal-tools-soy": "^6.0.0",
+    "ncp": "^2.0.0",
+    "node-sass": "^4.5.3",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.18.2",
+    "webpack": "^3.0.0",
+    "webpack-config-clay": "^2.3.4"
+  },
+  "browserslist": [
+    "extends browserslist-config-clay"
+  ]
 }
