+++ packages/clay-charts/src/scss/components/_arc.scss
