+++ packages/clay-charts/src/scss/components/_area.scss
