+++ packages/clay-charts/src/scss/components/_axis.scss
