+++ packages/clay-charts/src/scss/components/_bar.scss
