+++ packages/clay-charts/src/scss/components/_brush.scss
