+++ packages/clay-charts/src/scss/components/_chart.scss
