+++ packages/clay-charts/src/scss/components/_focus.scss
