+++ packages/clay-charts/src/scss/components/_legend.scss
