+++ packages/clay-charts/src/scss/components/_loading.scss
