+++ packages/clay-charts/src/scss/components/_point.scss
