+++ packages/clay-charts/src/scss/components/_select_drag.scss
