+++ packages/clay-charts/src/scss/components/_tooltip.scss
